+++ src/components/Poll/ValidInDaysSelector.tsx
@@ -1,6 +1,6 @@
 import { Popover, Transition } from '@headlessui/react';
 import { ChevronUpIcon } from '@heroicons/react/24/outline';
-import { t, Trans } from '@lingui/macro';
+import { plural, Plural, t, Trans } from '@lingui/macro';
 import { Fragment, useState } from 'react';
 
 import { ClickableButton } from '@/components/ClickableButton.js';
@@ -52,8 +52,8 @@ export function ValidInDaysSelector({ post: { poll }, readonly }: ValidInDaysSel
                         disabled={readonly}
                         className="flex h-6 items-center gap-2 rounded-full border border-lightMain px-3 text-[13px] font-bold text-lightMain"
                     >
-                        <span>{validInDays === 1 ? t`${validInDays} day` : t`${validInDays} days`}</span>
-                        {<ChevronUpIcon className={classNames('h-5 w-5 transition-all', open ? ' rotate-180' : '')} />}
+                        <Plural value={validInDays} _1={t`${validInDays} day`} other={`${validInDays} days`} />
+                        <ChevronUpIcon className={classNames('h-5 w-5 transition-all', { 'rotate-180': open })} />
                     </Popover.Button>
                     <Transition
                         as={Fragment}
